implemented attendance controller crud.

+++ src/main/java/com/dk/repository/AttendanceRepository.java
@@ -3,5 +3,5 @@ package com.dk.repository;
 import com.dk.model.Attendance;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AttendanceRepository extends CrudRepository<Attendance, Integer> {
+public interface AttendanceRepository extends CrudRepository<Attendance, Long> {
 }
